requirements_meshtastic_node.txt
paho-mqtt>=1.6.0
meshtastic>=2.7.3
protobuf>=6.0.0
cryptography>=3.0.0
PyYAML>=6.0

